pyproject.toml
[tool.poetry]
name = "django-video-streaming"
version = "0.1.0"
description = "Django Video Streaming Demo App with HLS Processing"
authors = ["Michael <michael@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
celery = {extras = ["redis"], version = "^5.3.4"}
redis = "^5.0.1"
psycopg2-binary = "^2.9.9"
python-decouple = "^3.8"
pillow = "^10.1.0"
django-filter = "^23.3"

[tool.poetry.group.dev.dependencies]
black = "^23.11.0"
flake8 = "^6.1.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "django"
line_length = 88
